Revert "получил первые фильмы с сервера"
This reverts commit 49880378

+++ app/src/main/java/ru/buryachenko/hw_look4films/recycler/ListFilmsAdapter.java
@@ -75,7 +75,7 @@ public class ListFilmsAdapter extends RecyclerView.Adapter<ListFilmsViewHolder> 
 
     private void callDetails(int position) {
         if (position != RecyclerView.NO_POSITION) {
-            Long previousSelected;
+            Integer previousSelected;
             if (!viewModel.getList().get(position).isSelected()) {
                 previousSelected = FilmInApp.getSelected();
                 if (previousSelected != null) {
